NIFI-7669 Moved time-based encryption tests to integration tests to avoid running during CI builds.
